requirements.txt
adal==1.2.7
applicationinsights==0.11.10
azure-common==1.1.27
azure-core==1.15.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.0.0
azure-mgmt-core==1.2.2
azure-mgmt-keyvault==9.0.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azureml-automl-core==1.31.0
azureml-automl-runtime==1.31.0
azureml-core==1.31.0
azureml-dataprep==2.18.0
azureml-dataprep-native==36.0.0
azureml-dataprep-rslex==1.16.1
azureml-dataset-runtime==1.31.0
azureml-defaults==1.31.0
azureml-interpret==1.31.0
azureml-model-management-sdk==1.0.1b6.post1
azureml-pipeline==1.31.0
azureml-pipeline-core==1.31.0
azureml-pipeline-steps==1.31.0
azureml-sdk==1.31.0
azureml-telemetry==1.31.0
azureml-train==1.31.0
azureml-train-automl-client==1.31.0
azureml-train-automl-runtime==1.31.0
azureml-train-core==1.31.0
azureml-train-restclients-hyperdrive==1.31.0
azure-storage-queue==12.1.6
backports.tempfile==1.0
backports.weakref==1.0.post1
boto==2.49.0
boto3==1.15.18
botocore==1.18.18
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
configparser==3.7.4
contextlib2==0.6.0
cryptography==3.4.7
Cython==0.29.14
dill==0.3.4
distro==1.5.0
docker==4.4.4
dotnetcore2==2.1.21
fire==0.4.0
Flask==1.0.3
flatbuffers==2.0
fusepy==3.0.1
gensim==3.8.3
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.6.0
interpret-community==0.17.2
interpret-core==0.2.4
isodate==0.6.0
itsdangerous==2.0.1
jeepney==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
json-logging-py==0.2
jsonpickle==2.0.0
keras2onnx==1.6.0
liac-arff==2.5.0
lightgbm==2.3.0
llvmlite==0.36.0
MarkupSafe==2.0.1
msal==1.12.0
msal-extensions==0.2.2
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
nimbusml==1.8.0
numba==0.53.1
numpy==1.18.5
oauthlib==3.1.1
onnx==1.7.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
onnxruntime==1.8.0
packaging==20.9
pandas==0.25.3
pathspec==0.8.1
patsy==0.5.1
pip==21.1.3
pipdeptree==2.0.0
pmdarima==1.1.1
portalocker==1.6.0
protobuf==3.17.3
psutil==5.8.0
pyarrow==3.0.0
pyasn1==0.4.8
pycparser==2.20
py-cpuinfo==5.0.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==227
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml.clib==0.2.4
ruamel.yaml==0.17.4
s3transfer==0.3.7
scikit-learn==0.22.2
scipy==1.5.2
SecretStorage==3.3.1
setuptools==47.1.0
shap==0.34.0
six==1.16.0
skl2onnx==1.4.9
sklearn-pandas==1.7.0
slicer==0.0.7
smart-open==1.9.0
statsmodels==0.10.2
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.25.11
websocket-client==1.1.0
Werkzeug==1.0.1
wfastcgi==3.0.0
wheel==0.36.2
zipp==3.4.1
